Add files via upload
This new one if for the 4th query on the manager side. It allows you to select from two dropdowns! one selects the location and the other selects what kind of position of employee you're looking for and returns the rest of their info. it took a few tries to get the 'All selection to work'